Put the engine data classes in separate file
